Change the Maven repositories, attaches sources and Javadocs and fixes POMs
This commit changes the Maven repositories to the new stable repository
of the SING group, both for publishing and for dependency retrieval.

In addition, the sources and Javadocs are attached to the deployment.
Due to this, the Javadoc of the source files have been reviewed an
corrected, avoiding errors during the Maven construction.

Finally the POM files have been reviewed, refactorized and some bugs have been
fixed. The main bug fixed was that the generated plugin's jars did not contain
the dependent libraries. Once fixed this bug, the configuration was also
changed to avoid the inclusion of testing libraries in the plugin's jars.