Fix AIBenchJTreeManager
An error was detected in the AIBenchJTreeManager, caused by the
datatypes that do not have an associated icon. The current code tried to
recover their icon as a global resource and, as the icon URL was null,
it caused an error.

This commit fixes this error by checking first whether the icon URL is
null or not.

In addition, this commit also simplifies some additional code in the
same class.