Reorganize /target layout
In order to make easy the building of the built application, a
/target/dist is generated containing those files needed to run the
application.

In /target remains the jar of the plugin, and some other maven generated
artifacts not needed for the application.

In addition, to make easy to place in the source tree those files that
will not go inside the plugin (such as /conf or the run scripts) a
/src/main/global-resources directory is created exclusively for these
kind of files.