Fix bugs in FileParamProvider
When the "selectionMode=filesAndDirectories" (or other non valid mode)
was used, a ArrayIndexOutOfBounds was thrown, as the 2 index was
accessed in an array with 2 values (property name and value).

On the other hand, the FileChooserConfiguration.configureFileChooser
method did not removed the actual filters of the FileChooser before
adding more, causing an unexpected behavior if the FileChooser already
has filters. The FileChooserConfiguration.clearFileChooser now removes
all the filters from the FileChooser.