Revise the FileParamProvider behavior and refactorize it
The FileParamProvider was designed so it kept the last directory
selected through the different FileParamProvider that could exist in an
application. However, this caused the file selection dialog to show the
last directory selected instead of the file actually selected in the
component in some cases. In addition, when the file selection mode
changed between different file providers, the file dialog showed an
empty selected file and prevented from selecting the last file selected
(this is, probably, a Swing bug).

Finally, the source code of the FileParamProvider has been refactorized
and simplified.