Fixes bugs: dialog's help icon location and port order
This commit fixes two bugs. The first one is related to the location of
the help icon used in the generated operation dialogs that it is used to
replace the description text, so that the text appears as a tooltip of
these icons. The code that replaces the description with this icons had
some bugs that caused the application to fail with certain
configurations.

The second bug was related to the port sorting comparator used in the
OperationDefinition class. The comparator didn't follow the comparator's
contract, causing the code to fail, for example, when to ports don't
have a specific port order.