Make FileParamProvider#setSelectedFile public
Make FileParamProvider#setSelectedFile public so that it can be
invoked, for example, from a custom params window that needs to
set the value of the selected file.

Moreover, setChanged and notifyObservers invocations have been
also moved to this method since it is the unique point when the
status changes. Previously, setChanged and notifyObservers where
being invoked after showing the file chooser, regardless of user
selection.