Add support for EnumParamProvider customization using extras
This commit adds support for EnumParamProvider customization using
extras in the corresponding Port annotation. This allows choosing
the type of component to display the enum constants: a combobox,
which is the default behaviour, or a set of radio buttons.

Moreover, this commit removes unused class EnumParamProvider and
renames EnumParamProviderCombo to EnumParamProvider.